Refactor start
